7-email-news-api/requirements.txt
requests == 2.28.2
python-dotenv == 1.0.0
